pyproject.toml
[project]
name = "repl-nix-flask-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.3",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.36",
]
